# arm64 crc prefetch optimise