@siying 2019-09-11
Our CI doesn't cover ARM and I have no platform to validate it. If someone voluntarily tries it out on ARM and make sure crc32_test passes, I can pass it.  CC @guyuqi @Zhiwei-Dai